requirements.txt
aiofiles==23.1.0
alembic==1.8.1
anyio==3.7.0
asyncpg==0.27.0
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.1.1   
click==8.1.3
colorama==0.4.6
cryptography==41.0.1        
databases==0.7.0
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.84.0
fastapi-users==11.0.0
fastapi-users-db-sqlalchemy==5.0.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.1
idna==3.4
iniconfig==2.0.0
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.3
orjson==3.8.0
packaging==23.1
passlib==1.7.4
pluggy==1.0.0
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
pytest==7.2.0
pytest-asyncio==0.20.2
python-dotenv==0.21.0
python-multipart==0.0.6
requests==2.28.1
rfc3986==1.5.0
shortuuid==1.0.11
sniffio==1.3.0
SQLAlchemy==2.0.16
SQLAlchemy-Utils==0.38.3
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.6.3
urllib3==1.26.16
uvicorn==0.18.3